src/backend/pyproject.toml
[project]
name = "drone-tm"
version = "2025.2.1"
description = "Community-driven drone imagery generation"
authors = [
    {name = "HOTOSM", email = "sysadmin@hotosm.org"},
]
dependencies = [
    "fastapi==0.112.0",
    "geojson==3.1.0",
    "uvicorn==0.30.5",
    "python-multipart>=0.0.9",
    "pydantic[email]>=2.8.2",
    "pydantic-settings==2.4.0",
    "geojson-pydantic==1.1.0",
    "shapely==2.0.5",
    "sqlalchemy==2.0.23",
    "geoalchemy2==0.14.2",
    "psycopg[pool]>=3.2.1",
    "requests>=2.32.3",
    "requests-oauthlib>=2.0.0",
    "loguru>=0.7.2",
    "minio>=7.2.7",
    "pyjwt>=2.8.0",
    "alembic>=1.13.1",
    "itsdangerous>=2.2.0",
    "Jinja2>=3.1.4",
    "numpy==1.26.4",
    "GDAL==3.6.2",
    "aiohttp>=3.11.11",
    "aiosmtplib>=3.0.1",
    "python-slugify>=8.0.4",
    "psycopg2>=2.9.9",
    "pyodm>=1.5.11",
    "asgiref>=3.8.1",
    "bcrypt>=4.2.1",
    "Scrapy==2.12.0",
    "asgi-lifespan>=2.1.0",
    "arq>=0.26.3",
    "redis>=5.2.1",
    "drone-flightplan",
]
requires-python = ">=3.11"
license = {text = "GPL-3.0-only"}

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
    "commitizen>=3.30.0",
]
test = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "httpx==0.27.2",
    "pyinstrument==5.0.0",
    "coverage>=7.6.4",
    "coverage-badge>=1.1.2",
]
debug = [
    "debugpy>=1.8.7",
]
docs = [
    "mkdocs==1.6.1",
    "mkdocs-material==9.5.42",
    "mkdocstrings-python==1.12.2",
    "mkdocs-exclude==1.0.2",
    "python-dotenv==1.0.1",
    "mkdocs-git-revision-date-localized-plugin>=1.3.0",
    "mdx-truly-sane-lists>=1.3",
]
monitoring = [
    "sentry-sdk[opentelemetry]>=2.17.0",
    #"opentelemetry-sdk>=1.27.0",
    #"opentelemetry-api>=1.27.0",
    #"opentelemetry-exporter-otlp>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-logging>=0.48b0",
    "opentelemetry-instrumentation-psycopg2>=0.48b0",
    "opentelemetry-instrumentation-requests>=0.48b0",
]

[tool.uv.sources]
drone-flightplan = { workspace = true }

[tool.uv.workspace]
members = ["packages/*"]

#[tool.ruff]
#fix = true
#line-length = 88
#target-version = "py311"
#exclude = [
#    ".ruff_cache",
#    "app/__version__.py",
#    "app/migrations",
#]
#[tool.ruff.lint]
#select = ["I", "E", "W", "D", "B", "F", "N", "Q"]
#[tool.ruff.lint.pydocstyle]
#convention = "google"
